+++ GameBoy/fram.h
@@ -32,5 +32,5 @@
 
 //FUNCTIONS
 void init_fram(void);
-int8_t write_fram(uint16_t address, byte *buf, byte count);
-int8_t read_fram(uint16_t address, byte *buf, byte count);
+int8_t write_fram(uint32_t address, byte *buf, byte count);
+int8_t read_fram(uint32_t address, byte *buf, byte count);
